requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.7.1
distro==1.9.0
fonttools==4.54.1
fpdf==1.7.2
fpdf2==2.8.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
openai==1.52.2
pdfminer==20191125
pdfminer.six==20240706
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
